+++ docs/src/releasenotes.md
@@ -3,6 +3,7 @@
 ## Unreleased
 * New unexported functions: `python_executable_path`, `python_library_path`, `python_library_handle` and `python_version`.
 * `Py` is now treated as a scalar when broadcasting.
+* `PyArray` is now serializable.
 * Bug fixes.
 
 ## 0.9.15 (2023-10-25)